First version of fcc_analyzer with PtQt4
The file is named: fcc_analyzer_PyQt4.py and lives with
the plain standard version

The currect version provides practically the same functionality
as compared with the standard version. For the moment the PyQt4
is only in develpment (not intalled with make).

The improvements, apart from the inclusion of a PyQt4 GUI include
a more pythonic coding style