Updates on fcc_analyzer_PyQt4: set Arial font in box
The font in the analysis box is set to Arial. Otherwise, the default
font in the compiling computer (Ubuntu) won't be present in other
systems, and the alternative font may provide different layouts
(e.g. splitting lines...). Arial should be present in all computers.

Other minor fixes, including the management to get the progression for
the cases when the even/odd values are not allowed