fcc_analyzer_PyQt4: support fort.18 instead of fort.22
If fort.22 is not present, fort.18 is used instead. In this
case, the convolution controls are not active.

Also, now, if the sript is called in a location where no fcclasses
output is present, then a diaolog is opened to specify the path.
This is useful when calling the program directly from an icon,
as in windows