Fix bug reading Turbomole atnames
Atom names are in low case in Turbomole, and the whole
name was turned to upper case. This is a problem for
two letters elements. Upper case NA, for instance can be
confusing as it can be N or Na in MM force fields.

Fixed, just changing the first letter.

At the same time, other modules were updated from mainstream:

 * fcc_basisc: assign_masses, updated from generic_io (molecular_tools)
               element_from_AtName: added from generic_io (molecular_tools)

 * constants:  atoms info handling from constants (molecular_tools)